# Faster implementation of maximum cardinality search.