--- a/Project.toml
+++ b/Project.toml
@@ -38,7 +38,7 @@ GraphRecipesExt = ["GraphRecipes", "Plots"]
 TikzGraphsExt = "TikzGraphs"
 
 [compat]
-CliqueTrees = "0.2.1"
+CliqueTrees = "0.3"
 Combinatorics = "1.0"
 DelimitedFiles = "1.6, 1.7, 1.8, 1.9"
 Distances = "0.8, 0.9, 0.10"