--- a/src/CausalInference.jl
+++ b/src/CausalInference.jl
@@ -8,7 +8,6 @@ using Combinatorics
 using Base.Iterators
 using Memoization, LRUCache
 using ThreadsX
-using LinkedLists
 
 import Base: ==, show