I have a library [CliqueTrees.jl](https://github.com/AlgebraicJulia/CliqueTrees.jl/tree/main) with a faster implementation of the maximum cardinality algorithm. Here are some benchmarks.

```julia
julia> using BenchmarkTools, MatrixMarket, SuiteSparseMatrixCollection, Graphs

julia> using CausalInference: count_mcs

julia> ssmc = ssmc_db(); name = "venturiLevel3";

julia> graph = DiGraph(mmread(joinpath(fetch_ssmc(ssmc[ssmc.name .== name, :], format="MM")[1], "$(name).mtx")))
{4026819, 16108474} directed simple Int64 graph
```

**current version**

```julia
julia> @benchmark count_mcs($graph)
BenchmarkTools.Trial: 6 samples with 1 evaluation per sample.
 Range (min … max):  710.482 ms …    1.613 s  ┊ GC (min … max): 32.27% … 69.52%
 Time  (median):     765.882 ms               ┊ GC (median):    32.35%
 Time  (mean ± σ):   895.824 ms ± 353.408 ms  ┊ GC (mean ± σ):  44.48% ± 15.14%

  █ ▁ ▁ ▁                                                     ▁  
  █▁█▁█▁█▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁█ ▁
  710 ms           Histogram: frequency by time          1.61 s <

 Memory estimate: 706.63 MiB, allocs estimate: 20134711.
```

**new version**

```julia
julia> @benchmark count_mcs($graph)
BenchmarkTools.Trial: 18 samples with 1 evaluation per sample.
 Range (min … max):  274.436 ms … 311.835 ms  ┊ GC (min … max): 0.00% … 1.62%
 Time  (median):     288.422 ms               ┊ GC (median):    1.17%
 Time  (mean ± σ):   289.501 ms ±  11.964 ms  ┊ GC (mean ± σ):  1.09% ± 1.10%

  ▁▁ █  ▁ ▁ ▁      ▁▁        ▁  ▁ ▁     █    ▁     ▁ ▁        ▁  
  ██▁█▁▁█▁█▁█▁▁▁▁▁▁██▁▁▁▁▁▁▁▁█▁▁█▁█▁▁▁▁▁█▁▁▁▁█▁▁▁▁▁█▁█▁▁▁▁▁▁▁▁█ ▁
  274 ms           Histogram: frequency by time          312 ms <

 Memory estimate: 153.61 MiB, allocs estimate: 15.
```

You can take a look at the implementation [here](https://github.com/AlgebraicJulia/CliqueTrees.jl/blob/5cafd6f69c7a2e6b489b06d9797c31285741e7eb/src/elimination_algorithms.jl#L391).