@mschauer 2025-02-15
Thank you for the PR! Looks very good, there is some conflict in the compat requirements of our packages, not sure what exactly conflicts. Cc @marcel3243