@samuelsonric 2025-02-15
> One could check whether this actually improves the performance of the Zigzag sampler. Maybe its not the bottleneck.
> 
> On another note: since coding this I noted that it's possible to implement MCS without linked lists. Just use a Vector for each of the sets and store the set that each vertex is currently is in (like is probably already done with something like `size[v]`).
> 
> To move vertices to a new set, just insert them in the new set (and don't delete from the old one). To retrieve the vertex with maximum cardinality, just pop vertices `v` from maximum cardinality set (say with cardinality `x`) until you find one with `size[v] = x`. I did something like this here: https://github.com/mwien/CliquePicking/blob/bfc95435174951c6ea8284260e1eda30d325aa11/cliquepicking_rs/src/chordal.rs#L3
> 
> Has same number of insertions/deletions as the linked list implementation. Memory consumption is limited by graph size. If I would reimplement MCS, I would probably take that route.

Since the sets are disjoint, you can store the whole collection using three vectors of length |V|. This reduces allocations, which improves performance.