@mwien 2025-02-15
> Since the sets are disjoint, you can store the whole collection using three vectors of length |V|. This reduces allocations, which improves performance.

Huh! Didn't know about this, cool idea :)