@samuelsonric 2025-02-15
> Thank you for the PR! Looks very good, there is some conflict in the compat requirements of our packages, not sure what exactly conflicts. Cc @marcel3243

I suspect that the problem is that I am only supporting Julia 1.10+, and you are testing with Julia 1.6. I will see if I can release an update that supports earlier versions of Julia.