@samuelsonric 2025-02-15
> One could check whether this actually improves the performance of the Zigzag sampler. Maybe its not the bottleneck.

I benchmarked the `causalzigzag(n; score, κ, iterations)` call in [this example](https://github.com/mschauer/CausalInference.jl/blob/1335e70390813d07d99611988be9c34fcf8c0d95/docs/src/examples/ges_basic_examples.md?plain=1#L66) on Julia 1.11.3.

**old version**

```julia
BenchmarkTools.Trial: 3 samples with 1 evaluation per sample.
 Range (min … max):  1.736 s …    1.963 s  ┊ GC (min … max): 15.96% … 24.19%
 Time  (median):     1.861 s               ┊ GC (median):    20.60%
 Time  (mean ± σ):   1.853 s ± 113.775 ms  ┊ GC (mean ± σ):  20.42% ±  4.13%

  █                               █                        █  
  █▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁█▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁█ ▁
  1.74 s         Histogram: frequency by time         1.96 s <

 Memory estimate: 2.99 GiB, allocs estimate: 60403438.
 ```
 
 **new version**
 
 ```julia
BenchmarkTools.Trial: 4 samples with 1 evaluation per sample.
 Range (min … max):  1.522 s …   1.669 s  ┊ GC (min … max): 25.69% … 30.39%
 Time  (median):     1.589 s              ┊ GC (median):    27.95%
 Time  (mean ± σ):   1.592 s ± 80.245 ms  ┊ GC (mean ± σ):  27.34% ±  3.69%

  █                                                 ▁     ▁  
  █▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁█▁▁▁▁▁█ ▁
  1.52 s         Histogram: frequency by time        1.67 s <

 Memory estimate: 3.06 GiB, allocs estimate: 60820094.
 ```
 
 It's hard to say, but the run-time seems to have decreased.