@samuelsonric 2025-02-15
I can push an update that supports Julia 1.8 and 1.9.

For 1.6 and 1.7, I would have to loosen some of the compatibility requirements in [TreeWidthSolver](https://github.com/ArrogantGao/TreeWidthSolver.jl).