@mschauer 2025-02-16
I am happy to drop < 1.8.