@samuelsonric 2025-02-16
Tests are passing on Julia versions
- v1.8.5
- v1.9.4
- v1.10.8
- v1.11.3