@samuelsonric 2025-02-16
CI failed because it's on Julia 1.7.