@mschauer 2025-02-16
Great, now CI is updated too. Is this ready to merge?